Move multi_transform to combine.py
