+++ optax/_src/wrappers_test.py
@@ -420,75 +420,5 @@ class MaybeUpdateTest(chex.TestCase):
       self.assertEqual(state.inner_state.found_nan, False)
 
 
-def _map_keys_fn(fn):
-  def map_fn(nested_dict):
-    return {k: (map_fn(v) if isinstance(v, dict) else fn(k, v))
-            for k, v in nested_dict.items()}
-  return map_fn
-
-
-class MultiTransformTest(chex.TestCase):
-  """Tests for the multi_transform wrapper."""
-
-  @chex.all_variants
-  @parameterized.parameters(True, False)
-  def test_multi_transform(self, use_fn):
-    params = {'a1': 1., 'b1': 2., 'z1': {'a2': 3., 'z2': {'c1': 4.}}}
-    params = jax.tree_map(jnp.asarray, params)
-    input_updates = jax.tree_map(lambda x: x / 10.0, params)
-    tx_dict = {'a': transform.scale(-1.0),
-               'b': transform.ema(0.0),  # stateful
-               'c': transform.scale(2.0)}
-    param_labels = _map_keys_fn(lambda k, _: k[0])
-    if not use_fn:
-      param_labels = param_labels(params)
-    tx = wrappers.multi_transform(tx_dict, param_labels)
-    update_fn = self.variant(tx.update)
-    state = self.variant(tx.init)(params)
-
-    correct_update_fn = _map_keys_fn(
-        lambda k, v: {'a': -v, 'b': v, 'c': 2.0*v}[k[0]])
-
-    updates, state = update_fn(input_updates, state, params)
-    correct_updates = correct_update_fn(input_updates)
-    chex.assert_tree_all_close(updates, correct_updates)
-
-    # Check repeated application, this time with no params.
-    correct_updates = correct_update_fn(correct_updates)
-    updates, state = update_fn(updates, state)
-    chex.assert_tree_all_close(updates, correct_updates)
-
-  @parameterized.parameters(list, tuple, dict)
-  def test_empty(self, container):
-    init_fn, update_fn = wrappers.multi_transform({0: _build_sgd()},
-                                                  lambda _: 0)
-    updates, _ = update_fn(container(), init_fn(container()))
-    self.assertEqual(updates, container())
-
-  @chex.all_variants
-  @parameterized.parameters(
-    (False, False), (False, True), (True, False), (True, True))
-  def test_labels_mismatch(self, use_extra_label, use_fn):
-    # The labels from label_fn must be a subet of the keys for the tx.
-    params = {'a': 1., 'b': [2., 3.], 'c': {'d': 4., 'e': (5., 6.)}}
-    params = jax.tree_map(jnp.asarray, params)
-    label_tree = {'a': 0, 'b': [1, 0], 'c': 1}  # prefix of params
-
-    if use_extra_label:
-      label_tree['a'] = 3
-
-    init_fn, update_fn = wrappers.multi_transform(
-        {0: _build_sgd(), 1: _build_simple_adam(), 2: transform.trace(1.0)},
-        (lambda _: label_tree) if use_fn else label_tree)
-
-    if use_extra_label:
-      with self.assertRaises(ValueError):
-        self.variant(init_fn)(params)
-    else:
-      state = self.variant(init_fn)(params)
-      updates = jax.tree_map(lambda x: x / 10.0, params)
-      self.variant(update_fn)(updates, state)
-
-
 if __name__ == '__main__':
   absltest.main()
